# openapi-specs/finance-api-docs.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "HMPPS Prisoner Finance Sync API",
    "description" : "API for synchronising NOMIS transactions with the Prisoner Finance Ledger.\n\n## Authentication\nThis API uses **OAuth2 with JWTs**. Pass your JWT in the `Authorization` header using the `Bearer` scheme.\n**Important:** This service is designed for client tokens only; user tokens should not be used.\n\n## Authorization\nAccess to API endpoints is controlled by roles. The required roles are documented with each endpoint.\nIntegrations should request one of the following roles based on their needs:\n\n* `ROLE_PRISONER_FINANCE_SYNC__RW`: Grants **read/write access** to the NOMIS sync endpoints.\n* `ROLE_PRISONER_FINANCE_SYNC__AUDIT__RO`: Grants **read only access** to the Sync audit endpoints.",
    "contact" : {
      "name" : "HMPPS Digital Studio",
      "email" : "feedback@digital.justice.gov.uk"
    },
    "version" : "2026-02-13.541.7e25a6e"
  },
  "servers" : [ {
    "url" : "https://prisoner-finance-sync-api-dev.hmpps.service.justice.gov.uk/",
    "description" : "Development"
  }, {
    "url" : "https://prisoner-finance-sync-api-preprod.hmpps.service.justice.gov.uk/",
    "description" : "Pre-Production"
  }, {
    "url" : "https://prisoner-finance-sync-api.hmpps.service.justice.gov.uk/",
    "description" : "Production"
  }, {
    "url" : "http://localhost:8080",
    "description" : "Local"
  } ],
  "security" : [ {
    "bearer-jwt" : [ ]
  } ],
...
  "paths" : {
...
    "/prisons/{prisonId}/accounts/{accountCode}/transactions/{transactionId}" : {
      "get" : {
        "tags" : [ "Prison accounts" ],
        "summary" : "Get a transaction for a specific account for a specific prison",
        "operationId" : "getPrisonAccountTransaction",
        "parameters" : [ {
          "name" : "prisonId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "pattern" : "^[A-Za-z0-9]{3}$"
          }
        }, {
          "name" : "accountCode",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "transactionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TransactionDetailsList"
                }
              }
            }
          },
          "404" : {
            "description" : "Prison account not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ "ROLE_PRISONER_FINANCE_SYNC__RW" ]
        } ]
      }
    },
...
    "/prisoners/{prisonNumber}/accounts/{accountCode}/transactions/{transactionId}" : {
      "get" : {
        "tags" : [ "Prisoner trust accounts" ],
        "summary" : "Get a transaction for a specific sub account for a specific prisoner",
        "operationId" : "getPrisonerAccountTransaction",
        "parameters" : [ {
          "name" : "prisonNumber",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "accountCode",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "transactionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TransactionDetailsList"
                }
              }
            }
          },
          "404" : {
            "description" : "Prisoner account not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-jwt" : [ "ROLE_PRISONER_FINANCE_SYNC__RW" ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ErrorResponse" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "errorCode" : {
            "type" : "string"
          },
          "userMessage" : {
            "type" : "string"
          },
          "developerMessage" : {
            "type" : "string"
          },
          "moreInfo" : {
            "type" : "string"
          }
        },
        "required" : [ "status" ]
      },
...
      "TransactionAccountDetails" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "name" : {
            "type" : "string"
          },
          "transactionType" : {
            "type" : "string"
          },
          "transactionDescription" : {
            "type" : "string"
          },
          "prison" : {
            "type" : "string"
          },
          "prisoner" : {
            "type" : "string"
          },
          "classification" : {
            "type" : "string"
          },
          "postingType" : {
            "type" : "string",
            "enum" : [ "DR", "CR" ]
          }
        },
        "required" : [ "classification", "code", "name", "postingType" ]
      },
      "TransactionDetails" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "date" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "reference" : {
            "type" : "string"
          },
          "clientRequestId" : {
            "type" : "string"
          },
          "postings" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TransactionPosting"
            }
          }
        },
        "required" : [ "date", "id", "postings", "type" ]
      },
      "TransactionDetailsList" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TransactionDetails"
            }
          }
        },
        "required" : [ "items" ]
      },
      "TransactionPosting" : {
        "type" : "object",
        "properties" : {
          "account" : {
            "$ref" : "#/components/schemas/TransactionAccountDetails"
          },
          "address" : {
            "type" : "string"
          },
          "postingType" : {
            "type" : "string",
            "enum" : [ "DR", "CR" ]
          },
          "amount" : {
            "type" : "number"
          }
        },
        "required" : [ "amount", "postingType" ]
      },
...
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "type" : "http",
        "description" : "A HMPPS Auth access token with the `ROLE_PRISONER_FINANCE_SYNC__RW` role.",
        "name" : "Authorization",
        "in" : "header",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}